+++ source/module_base/test/data/cos_periodic.dat
@@ -1,0 +1,7 @@
+periodic periodic
+ 0.00000000000000000e+00   1.00000000000000006e-01   1.00000000000000000e+00   1.50000000000000000e+00   3.00000000000000000e+00   6.28318530717958623e+00  
+ 1.00000000000000000e+00   9.95004165278025821e-01   5.40302305868139765e-01   7.07372016677029064e-02  -9.89992496600445415e-01   1.00000000000000000e+00  
+ 0.00000000000000000e+00   1.74532925199432948e-01   3.49065850398865896e-01   5.23598775598298816e-01   6.98131700797731791e-01   8.72664625997164767e-01   1.04719755119659763e+00   1.22173047639603061e+00   1.39626340159546358e+00   1.57079632679489656e+00   1.74532925199432953e+00   1.91986217719376251e+00   2.09439510239319526e+00   2.26892802759262846e+00   2.44346095279206121e+00   2.61799387799149441e+00   2.79252680319092716e+00   2.96705972839035992e+00   3.14159265358979312e+00   3.31612557878922587e+00   3.49065850398865907e+00   3.66519142918809182e+00   3.83972435438752502e+00   4.01425727958695777e+00   4.18879020478639053e+00   4.36332312998582328e+00   4.53785605518525692e+00   4.71238898038468967e+00   4.88692190558412243e+00   5.06145483078355518e+00   5.23598775598298882e+00   5.41052068118242158e+00   5.58505360638185433e+00   5.75958653158128708e+00   5.93411945678071984e+00   6.10865238198015348e+00   6.28318530717958623e+00  
+ 1.00000000000000000e+00   9.84381299886306516e-01   9.38016522020507670e-01   8.63713424115361450e-01   7.64165387715657451e-01   6.42065794366185383e-01   5.00118352188328785e-01   3.42056178626561569e-01   1.73503016854116870e-01   2.70592234453568648e-04  -1.72386293431320586e-01  -3.39925662651143046e-01  -4.97852121855354546e-01  -6.41670277474296391e-01  -7.66884735938309081e-01  -8.69000103677733948e-01  -9.43520987122911547e-01  -9.85951992704182878e-01  -9.92542393974230608e-01  -9.64850207988183883e-01  -9.06738697819773454e-01  -8.22080502627916765e-01  -7.14748261571530152e-01  -5.88614613809530951e-01  -4.47552198500836496e-01  -2.95433654804362900e-01  -1.36131621879026887e-01   2.64812611162533740e-02   1.88532355022561493e-01   3.46149020680980524e-01   4.95458618932594463e-01   6.32588510618484978e-01   7.53666056579735066e-01   8.54818617657429058e-01   9.32173554692648398e-01   9.81858228526478083e-01   1.00000000000000000e+00  
+ 6.18473891098361445e-03  -1.80469237892337431e-01  -3.48259990143118536e-01  -5.00618802880412384e-01  -6.37545676104219083e-01  -7.59040609814538803e-01  -8.64447219749984619e-01  -9.41247940672875094e-01  -9.84665089228650259e-01  -9.95326120656199964e-01  -9.78883265201603403e-01  -9.36683701867835450e-01  -8.68727430654895882e-01  -7.75014451562784812e-01  -6.55544764591502682e-01  -5.10318369741048716e-01  -3.39335267011423580e-01  -1.42595456402626830e-01   6.41233977052624166e-02   2.49507580803381257e-01   4.12703175963197166e-01   5.53710183184709215e-01   6.72528602467918124e-01   7.69158433812823450e-01   8.43599677219425303e-01   8.95852332687724018e-01   9.25916400217719260e-01   9.33791879809410919e-01   9.19478771462799216e-01   8.82977075177883819e-01   8.24286790954665616e-01   7.43407918793143718e-01   6.40340458693318570e-01   5.15084410655190172e-01   3.67639774678758080e-01   1.98006550764022293e-01   6.18473891098361445e-03  
+-1.16262364588415146e+00  -1.00557944472069671e+00  -9.17160944338298312e-01  -8.28742443955899910e-01  -7.40323943573501397e-01  -6.51905443191102996e-01  -5.35672607329632511e-01  -3.44398827159106469e-01  -1.53125046988580316e-01   2.04230892878108500e-02   1.67998154850598658e-01   3.15573220413386479e-01   4.63148285976174190e-01   6.10723351538962067e-01   7.58298417101749722e-01   9.05873482664537821e-01   1.05344854822732525e+00   1.20102361379011291e+00   1.12573875011125457e+00   9.98607505889287705e-01   8.71476261667320506e-01   7.44345017445353752e-01   6.17213773223386442e-01   4.90082529001419687e-01   3.62951284779452932e-01   2.35820040557485955e-01   1.08688796335518534e-01  -1.84424478864484431e-02  -1.45573692108415198e-01  -2.72704936330382175e-01  -3.99836180552349374e-01  -5.26967424774316351e-01  -6.54098668996283328e-01  -7.81229913218249861e-01  -9.08361157440217282e-01  -1.03549240166218426e+00  -1.16262364588415146e+00
