+++ source/module_base/test/data/exp_first_deriv.dat
@@ -1,0 +1,7 @@
+first_deriv:1.10517091807564771e+00  first_deriv:2.71828182845904509e+00  
+ 1.00000000000000006e-01   1.29154966501488389e-01   1.66810053720005874e-01   2.15443469003188337e-01   2.78255940220712428e-01   3.59381366380462752e-01   4.64158883361277863e-01   5.99484250318940926e-01   7.74263682681126997e-01   1.00000000000000000e+00  
+ 1.10517091807564771e+00   1.13786644168535322e+00   1.18152981679905356e+00   1.24041185922219421e+00   1.32082420599103290e+00   1.43244298301307027e+00   1.59067568226275657e+00   1.82117928549843455e+00   2.16899447087117636e+00   2.71828182845904509e+00  
+ 1.00000000000000006e-01   1.25000000000000000e-01   1.50000000000000022e-01   1.75000000000000017e-01   2.00000000000000011e-01   2.25000000000000006e-01   2.50000000000000000e-01   2.75000000000000022e-01   3.00000000000000044e-01   3.25000000000000011e-01   3.49999999999999978e-01   3.75000000000000000e-01   4.00000000000000022e-01   4.25000000000000044e-01   4.50000000000000067e-01   4.74999999999999978e-01   5.00000000000000000e-01   5.25000000000000022e-01   5.50000000000000044e-01   5.75000000000000067e-01   5.99999999999999978e-01   6.25000000000000000e-01   6.50000000000000022e-01   6.75000000000000044e-01   7.00000000000000067e-01   7.24999999999999978e-01   7.50000000000000000e-01   7.75000000000000022e-01   8.00000000000000044e-01   8.25000000000000067e-01   8.49999999999999978e-01   8.75000000000000000e-01   9.00000000000000022e-01   9.25000000000000044e-01   9.50000000000000067e-01   9.74999999999999978e-01   1.00000000000000000e+00  
+ 1.10517091807564771e+00   1.13314845305620615e+00   1.16183423781052575e+00   1.19124621006218145e+00   1.22140274895196677e+00   1.25232269983173161e+00   1.28402537693915098e+00   1.31653067826587877e+00   1.34985870568413047e+00   1.38403052288246253e+00   1.41906754700129101e+00   1.45499127838897557e+00   1.49182434073154035e+00   1.52959016783122848e+00   1.56831221140395694e+00   1.60801396407294228e+00   1.64872023299870918e+00   1.69045738865219253e+00   1.73325189263525137e+00   1.77713020654974518e+00   1.82211879200369298e+00   1.86824485646834759e+00   1.91553841190619289e+00   1.96403012931547272e+00   2.01375067969443089e+00   2.06473073404130947e+00   2.11700096335435450e+00   2.17059203866204742e+00   2.22553592214226725e+00   2.28186930422749290e+00   2.33962995740218638e+00   2.39885565415081015e+00   2.45958416695782534e+00   2.52185326830769485e+00   2.58570073068487982e+00   2.65116432657384271e+00   2.71828182845904465e+00  
+ 1.10517091807564771e+00   1.13314857626511545e+00   1.16183448755634799e+00   1.19124518701385762e+00   1.22140376069562406e+00   1.25232046400808739e+00   1.28402663663421768e+00   1.31653040155273815e+00   1.34985411865053440e+00   1.38403404177963485e+00   1.41907067229022621e+00   1.45497988810568324e+00   1.49182144407227635e+00   1.52960106867554990e+00   1.56831876191550346e+00   1.60798584383446830e+00   1.64869208090593933e+00   1.69045678205178129e+00   1.73327994727199441e+00   1.77716157656657869e+00   1.82210170576561081e+00   1.86818792472837214e+00   1.91550098362087740e+00   1.96404088244312680e+00   2.01380762119512058e+00   2.06480119987685740e+00   2.11702161848833859e+00   2.17046900023482614e+00   2.22529349474582894e+00   2.28162888863530977e+00   2.33947518190326731e+00   2.39883237454970333e+00   2.45970046657461650e+00   2.52207945797800726e+00   2.58596934875987605e+00   2.65137013892022200e+00   2.71828182845904553e+00  
+ 1.10510282004738825e+00   1.13310983511001884e+00   1.16189187653084680e+00   1.19124256407428741e+00   1.22144333046701914e+00   1.25229505919741335e+00   1.28419875089301683e+00   1.31610244258862052e+00   1.35006277753420334e+00   1.38433107279383738e+00   1.41859936805347142e+00   1.45490086593012635e+00   1.49242361139733326e+00   1.52994635686453995e+00   1.56746910233174686e+00   1.60708020137141450e+00   1.64941876434625834e+00   1.69175732732110240e+00   1.73409589029594624e+00   1.77643445327079030e+00   1.81891195991557297e+00   1.86798555710533210e+00   1.91705915429509122e+00   1.96613275148485056e+00   2.01520634867460968e+00   2.06427994586436858e+00   2.11335354305412793e+00   2.16276179287056580e+00   2.22319776800966951e+00   2.28363374314877277e+00   2.34406971828787603e+00   2.40450569342697928e+00   2.46494166856608299e+00   2.52537764370518625e+00   2.58581361884428951e+00   2.64624959398339277e+00   2.70668556912249603e+00
