+++ source/module_base/test/data/log_second_deriv.dat
@@ -1,0 +1,7 @@
+second_deriv:-1.00000000000000000e+00 second_deriv:-1.00000000000000002e-02 
+ 1.00000000000000000e+00   1.29154966501488389e+00   1.66810053720005880e+00   2.15443469003188381e+00   2.78255940220712450e+00   3.59381366380462763e+00   4.64158883361277841e+00   5.99484250318940859e+00   7.74263682681126930e+00   1.00000000000000000e+01  
+ 0.00000000000000000e+00   2.55842788110449526e-01   5.11685576220899052e-01   7.67528364331348634e-01   1.02337115244179810e+00   1.27921394055224780e+00   1.53505672866269705e+00   1.79089951677314629e+00   2.04674230488359621e+00   2.30258509299404590e+00  
+ 1.00000000000000000e+00   1.25000000000000000e+00   1.50000000000000000e+00   1.75000000000000000e+00   2.00000000000000000e+00   2.25000000000000000e+00   2.50000000000000000e+00   2.75000000000000000e+00   3.00000000000000000e+00   3.25000000000000000e+00   3.50000000000000000e+00   3.75000000000000000e+00   4.00000000000000000e+00   4.25000000000000000e+00   4.50000000000000000e+00   4.75000000000000000e+00   5.00000000000000000e+00   5.25000000000000000e+00   5.50000000000000000e+00   5.75000000000000000e+00   6.00000000000000000e+00   6.25000000000000000e+00   6.50000000000000000e+00   6.75000000000000000e+00   7.00000000000000000e+00   7.25000000000000000e+00   7.50000000000000000e+00   7.75000000000000000e+00   8.00000000000000000e+00   8.25000000000000000e+00   8.50000000000000000e+00   8.75000000000000000e+00   9.00000000000000000e+00   9.25000000000000000e+00   9.50000000000000000e+00   9.75000000000000000e+00   1.00000000000000000e+01  
+ 0.00000000000000000e+00   2.23201685990528503e-01   4.05482161721055334e-01   5.59647381157334922e-01   6.93212288502453533e-01   8.10938183318088091e-01   9.16348065656249267e-01   1.01160453727708877e+00   1.09865227971226798e+00   1.17872100366669685e+00   1.25277767969836296e+00   1.32176777395878364e+00   1.38635002079692837e+00   1.44697637249515854e+00   1.50409420887274625e+00   1.55814764821316754e+00   1.60947600084362952e+00   1.65829393447507156e+00   1.70480885095700785e+00   1.74922815213895189e+00   1.79175923970966466e+00   1.83259005049654222e+00   1.87183533226151133e+00   1.90959263384970646e+00   1.94595950410626362e+00   1.98103349187631750e+00   2.01491214600500346e+00   2.04769301510038781e+00   2.07946352550178926e+00   2.11027403528676105e+00   2.14016641953235709e+00   2.16918255331563126e+00   2.19736431171363611e+00   2.22475356980342642e+00   2.25139220266205564e+00   2.27732208536657810e+00   2.30258509299404635e+00  
+ 1.00254261613056794e+00   7.98334999625206598e-01   6.66752209256807338e-01   5.71952135021786212e-01   4.99597454958138398e-01   4.44699888470842797e-01   3.99955939076224509e-01   3.63472602732268946e-01   3.33587746439520605e-01   3.07606422584049077e-01   2.85491363057420178e-01   2.66837284058103641e-01   2.50118943859901521e-01   2.35190122938787882e-01   2.22050821294762640e-01   2.10610784319237826e-01   2.00154304624132307e-01   1.90527432327081331e-01   1.81730167428084816e-01   1.73762509927142816e-01   1.66624366318174705e-01   1.60087152540269423e-01   1.53940134142904944e-01   1.48183311126081213e-01   1.42816683489798313e-01   1.37840251234056188e-01   1.33254014358854894e-01   1.29057876274537020e-01   1.25134123654711077e-01   1.21377871343099950e-01   1.17789119339703638e-01   1.14367867644522142e-01   1.11114116257555462e-01   1.08027865178803598e-01   1.05109114408266549e-01   1.02357863945944316e-01   9.97741137918368992e-02  
+-1.00000000000000377e+00  -6.33660932042887648e-01  -4.49235132334255161e-01  -3.25782694882304436e-01  -2.53054745626878019e-01  -1.95497023679798904e-01  -1.62454571477147702e-01  -1.29412119274496529e-01  -1.11657824079571477e-01  -9.61927667642009099e-02  -8.07277094488303704e-02  -7.04523993469851478e-02  -6.32943222386314758e-02  -5.61362451302777968e-02  -4.89781680219241247e-02  -4.34851335765309294e-02  -4.01667039843129556e-02  -3.68482743920949818e-02  -3.35298447998770149e-02  -3.02114152076590411e-02  -2.69292458727027513e-02  -2.53684643505395734e-02  -2.38076828283763989e-02  -2.22469013062132209e-02  -2.06861197840500430e-02  -1.91253382618868650e-02  -1.75645567397236871e-02  -1.60300110957334006e-02  -1.53600098628741376e-02  -1.46900086300148763e-02  -1.40200073971556132e-02  -1.33500061642963519e-02  -1.26800049314370888e-02  -1.20100036985778258e-02  -1.13400024657185645e-02  -1.06700012328593032e-02  -1.00000000000000401e-02
