+++ source/module_base/test/data/sin_not_a_knot.dat
@@ -1,0 +1,7 @@
+not-a-knot not-a-knot
+ 1.00000000000000006e-01   1.29154966501488389e-01   1.66810053720005874e-01   2.15443469003188337e-01   2.78255940220712428e-01   3.59381366380462752e-01   4.64158883361277863e-01   5.99484250318940926e-01   7.74263682681126997e-01   1.00000000000000000e+00  
+ 9.98334166468281548e-02   1.28796193418703991e-01   1.66037531159675150e-01   2.13780666055298912e-01   2.74679090976688467e-01   3.51695188663471381e-01   4.47670834718957189e-01   5.64216731736942645e-01   6.99189845133651677e-01   8.41470984807896505e-01  
+ 1.00000000000000006e-01   1.25000000000000000e-01   1.50000000000000022e-01   1.75000000000000017e-01   2.00000000000000011e-01   2.25000000000000006e-01   2.50000000000000000e-01   2.75000000000000022e-01   3.00000000000000044e-01   3.25000000000000011e-01   3.49999999999999978e-01   3.75000000000000000e-01   4.00000000000000022e-01   4.25000000000000044e-01   4.50000000000000067e-01   4.74999999999999978e-01   5.00000000000000000e-01   5.25000000000000022e-01   5.50000000000000044e-01   5.75000000000000067e-01   5.99999999999999978e-01   6.25000000000000000e-01   6.50000000000000022e-01   6.75000000000000044e-01   7.00000000000000067e-01   7.24999999999999978e-01   7.50000000000000000e-01   7.75000000000000022e-01   8.00000000000000044e-01   8.25000000000000067e-01   8.49999999999999978e-01   8.75000000000000000e-01   9.00000000000000022e-01   9.25000000000000044e-01   9.50000000000000067e-01   9.74999999999999978e-01   1.00000000000000000e+00  
+ 9.98334166468281548e-02   1.24674734514712698e-01   1.49438130358688637e-01   1.74108137455688206e-01   1.98669331470162691e-01   2.23106356655036864e-01   2.47403944178155422e-01   2.71546935540866008e-01   2.95520208213219726e-01   3.19308806225840347e-01   3.42897834817652569e-01   3.66272425041630545e-01   3.89418057463319811e-01   4.12320464705872358e-01   4.34965384966057578e-01   4.57338563261367681e-01   4.79425963790102938e-01   5.01213811410824883e-01   5.22688346176908625e-01   5.43835808141729826e-01   5.64642437361874472e-01   5.85094862720930697e-01   6.05181175115716008e-01   6.24889809005897301e-01   6.44209198851141918e-01   6.63127779111117310e-01   6.81633984245490487e-01   6.99716248713929012e-01   7.17363006976099671e-01   7.34562693491669916e-01   7.51303742720306977e-01   7.67574589121678086e-01   7.83363667155450361e-01   7.98659411281291143e-01   8.13450255958867552e-01   8.27724635647847040e-01   8.41470984807896505e-01  
+ 9.95004923184505952e-01   9.92197390241733368e-01   9.88771161271814858e-01   9.84726484485526354e-01   9.80066691602079154e-01   9.74793416196813478e-01   9.68912581684626351e-01   9.62425723364175201e-01   9.55336713363020729e-01   9.47651829867646733e-01   9.39371159698343727e-01   9.30499643073893545e-01   9.21055871973040285e-01   9.11041628742959020e-01   9.00456913383649749e-01   8.89303613354681888e-01   8.77596695966640139e-01   8.65339380713611495e-01   8.52531667595595843e-01   8.39173556612593186e-01   8.25265066443230833e-01   8.10851858679044746e-01   7.95976029301547761e-01   7.80637578310739877e-01   7.64836505706621206e-01   7.48572811489191747e-01   7.31846495658451279e-01   7.14657558214399802e-01   6.97005999157037537e-01   6.78891818486364262e-01   6.60315016202380312e-01   6.41275592305085462e-01   6.21773546794479715e-01   6.01808879670563179e-01   5.81381590933335857e-01   5.60491680582797747e-01   5.39139148618948627e-01  
+-9.99273971679864453e-02  -1.24675238253821624e-01  -1.49423079339714465e-01  -1.74110306895194078e-01  -1.98673123780586686e-01  -2.23112904322202821e-01  -2.47353856652764159e-01  -2.71594808983325553e-01  -2.95479606336378597e-01  -3.19311073293541137e-01  -3.43142540250703731e-01  -3.66341401449573223e-01  -3.89160286618692008e-01  -4.11979171787810738e-01  -4.34798056956929524e-01  -4.57268738221928150e-01  -4.79284652821407420e-01  -5.01300567420886689e-01  -5.23316482020365958e-01  -5.45332396619845228e-01  -5.67275878301223169e-01  -5.85780742833658996e-01  -6.04285607366094824e-01  -6.22790471898530651e-01  -6.41295336430966478e-01  -6.59800200963402195e-01  -6.78305065495838022e-01  -6.96809930028275515e-01  -7.15314794560709233e-01  -7.33819659093142951e-01  -7.52324523625576669e-01  -7.70829388158010387e-01  -7.89334252690444216e-01  -8.07839117222877934e-01  -8.26343981755311652e-01  -8.44848846287745370e-01  -8.63353710820179088e-01
