+++ source/module_base/test/data/sqrt_mix_bc.dat
@@ -1,0 +1,7 @@
+first_deriv:5.00000000000000000e-01  not-a-knot 
+ 1.00000000000000000e+00   1.29154966501488389e+00   1.66810053720005880e+00   2.15443469003188381e+00   2.78255940220712450e+00   3.59381366380462763e+00   4.64158883361277841e+00   5.99484250318940859e+00   7.74263682681126930e+00   1.00000000000000000e+01  
+ 1.00000000000000000e+00   1.13646366638572482e+00   1.29154966501488389e+00   1.46779926762206947e+00   1.66810053720005880e+00   1.89573565240637598e+00   2.15443469003188381e+00   2.44843674682222678e+00   2.78255940220712450e+00   3.16227766016837952e+00  
+ 1.00000000000000000e+00   1.25000000000000000e+00   1.50000000000000000e+00   1.75000000000000000e+00   2.00000000000000000e+00   2.25000000000000000e+00   2.50000000000000000e+00   2.75000000000000000e+00   3.00000000000000000e+00   3.25000000000000000e+00   3.50000000000000000e+00   3.75000000000000000e+00   4.00000000000000000e+00   4.25000000000000000e+00   4.50000000000000000e+00   4.75000000000000000e+00   5.00000000000000000e+00   5.25000000000000000e+00   5.50000000000000000e+00   5.75000000000000000e+00   6.00000000000000000e+00   6.25000000000000000e+00   6.50000000000000000e+00   6.75000000000000000e+00   7.00000000000000000e+00   7.25000000000000000e+00   7.50000000000000000e+00   7.75000000000000000e+00   8.00000000000000000e+00   8.25000000000000000e+00   8.50000000000000000e+00   8.75000000000000000e+00   9.00000000000000000e+00   9.25000000000000000e+00   9.50000000000000000e+00   9.75000000000000000e+00   1.00000000000000000e+01  
+ 1.00000000000000000e+00   1.11803728698643989e+00   1.22475682769484950e+00   1.32287978023881414e+00   1.41422464207946574e+00   1.50000375623387994e+00   1.58115482205509994e+00   1.65831336044919975e+00   1.73206149587978508e+00   1.80279292004163660e+00   1.87083207865261780e+00   1.93649703530829753e+00   2.00002057410239775e+00   2.06157397820173083e+00   2.12132717083545463e+00   2.17944902427163800e+00   2.23607463862589828e+00   2.29129895056073130e+00   2.34521455546809943e+00   2.39791404873996372e+00   2.44949002569315066e+00   2.50002598372984819e+00   2.54957121153853006e+00   2.59816695899999894e+00   2.64585447599505841e+00   2.69267501240451246e+00   2.73866981810916421e+00   2.78388014298981767e+00   2.82834723692727730e+00   2.87211234980234575e+00   2.91521673149582661e+00   2.95770163188852342e+00   2.99960830086124020e+00   3.04097798829478050e+00   3.08185194406994789e+00   3.12227141806754682e+00   3.16227766016837952e+00  
+ 5.00000000000000000e-01   4.47098909750842544e-01   4.08224810882382050e-01   3.78053779833561365e-01   3.53481662512829198e-01   3.33407818352801533e-01   3.16209958323798612e-01   3.01467598935841907e-01   2.88742679905178157e-01   2.77324939467649523e-01   2.67204555498214813e-01   2.58260988871481856e-01   2.50040603634092751e-01   2.42499911313340782e-01   2.35638911909225951e-01   2.29428522778511251e-01   2.23638122316879184e-01   2.18218103423094306e-01   2.13168466097156617e-01   2.08489210339066144e-01   2.04180292443896155e-01   2.00134871770223183e-01   1.96254450619764736e-01   1.92539028992520816e-01   1.88988606888491395e-01   1.85603184307676500e-01   1.82382761250076131e-01   1.79327337715690260e-01   1.76436913704518944e-01   1.73711489216562182e-01   1.71151064251819918e-01   1.68755638810292180e-01   1.66525212891979024e-01   1.64459786496880367e-01   1.62559359624996236e-01   1.60823932276326687e-01   1.59253504450871608e-01  
+-2.45211727308500382e-01  -1.77996994684759102e-01  -1.37423541149741701e-01  -1.07607040737062801e-01  -8.89698978287946640e-02  -7.37024413981044041e-02  -6.38804388339192647e-02  -5.40584362697341184e-02  -4.82656746863025995e-02  -4.30762488139266458e-02  -3.78868229415506921e-02  -3.42409267828305136e-02  -3.15221551162820982e-02  -2.88033834497336862e-02  -2.60846117831852707e-02  -2.39023649822226771e-02  -2.24208387108338972e-02  -2.09393124394451208e-02  -1.94577861680563409e-02  -1.79762598966675645e-02  -1.65116817411209045e-02  -1.58516836482628341e-02  -1.51916855554047638e-02  -1.45316874625466934e-02  -1.38716893696886247e-02  -1.32116912768305544e-02  -1.25516931839724857e-02  -1.18916950911143095e-02  -1.12316969982561819e-02  -1.05716989053980526e-02  -9.91170081253992322e-03  -9.25170271968179560e-03  -8.59170462682366798e-03  -7.93170653396553862e-03  -7.27170844110741014e-03  -6.61171034824928165e-03  -5.95171225539115403e-03
