+++ source/module_base/test/data/three_points_not_a_knot.dat
@@ -1,0 +1,7 @@
+not-a-knot not-a-knot
+ 0.00000000000000000e+00   2.99999999999999989e-01   1.00000000000000000e+00  
+ 0.00000000000000000e+00   2.95520206661339546e-01   8.41470984807896505e-01  
+ 0.00000000000000000e+00   2.77777777777777762e-02   5.55555555555555525e-02   8.33333333333333287e-02   1.11111111111111105e-01   1.38888888888888895e-01   1.66666666666666657e-01   1.94444444444444420e-01   2.22222222222222210e-01   2.50000000000000000e-01   2.77777777777777790e-01   3.05555555555555525e-01   3.33333333333333315e-01   3.61111111111111105e-01   3.88888888888888840e-01   4.16666666666666630e-01   4.44444444444444420e-01   4.72222222222222210e-01   5.00000000000000000e-01   5.27777777777777790e-01   5.55555555555555580e-01   5.83333333333333259e-01   6.11111111111111049e-01   6.38888888888888840e-01   6.66666666666666630e-01   6.94444444444444420e-01   7.22222222222222210e-01   7.50000000000000000e-01   7.77777777777777679e-01   8.05555555555555469e-01   8.33333333333333259e-01   8.61111111111111049e-01   8.88888888888888840e-01   9.16666666666666630e-01   9.44444444444444420e-01   9.72222222222222210e-01   1.00000000000000000e+00  
+ 0.00000000000000000e+00   2.89141774610951995e-02   5.75117844399816239e-02   8.57928209366592559e-02   1.13757286951128120e-01   1.41405182483388209e-01   1.68736507533439495e-01   1.95751262101282020e-01   2.22449446186915784e-01   2.48831059790340758e-01   2.74896102911556972e-01   3.00644575550564286e-01   3.26076477707362977e-01   3.51191809381952769e-01   3.75990570574333827e-01   4.00472761284506151e-01   4.24638381512469687e-01   4.48487431258224378e-01   4.72019910521770392e-01   4.95235819303107561e-01   5.18135157602235941e-01   5.40717925419155421e-01   5.62984122753866445e-01   5.84933749606368458e-01   6.06566805976661683e-01   6.27883291864746340e-01   6.48883207270621987e-01   6.69566552194288955e-01   6.89933326635747024e-01   7.09983530594996526e-01   7.29717164072037017e-01   7.49134227066868941e-01   7.68234719579492076e-01   7.87018641609906422e-01   8.05485993158111757e-01   8.23636774224108636e-01   8.41470984807896394e-01  
+ 1.04660865727918528e+00   1.03521211991966933e+00   1.02381558256015315e+00   1.01241904520063719e+00   1.00102250784112101e+00   9.89625970481604944e-01   9.78229433122088876e-01   9.66832895762572919e-01   9.55436358403056851e-01   9.44039821043540783e-01   9.32643283684024715e-01   9.21246746324508647e-01   9.09850208964992579e-01   8.98453671605476623e-01   8.87057134245960555e-01   8.75660596886444487e-01   8.64264059526928530e-01   8.52867522167412462e-01   8.41470984807896394e-01   8.30074447448380437e-01   8.18677910088864369e-01   8.07281372729348412e-01   7.95884835369832344e-01   7.84488298010316276e-01   7.73091760650800319e-01   7.61695223291284251e-01   7.50298685931768183e-01   7.38902148572252226e-01   7.27505611212736158e-01   7.16109073853220091e-01   7.04712536493704134e-01   6.93315999134188066e-01   6.81919461774672109e-01   6.70522924415156041e-01   6.59126387055640084e-01   6.47729849696123905e-01   6.36333312336607948e-01  
+-4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942578168e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01  -4.10275344942577003e-01
