+++ source/module_base/test/data/two_points_first_deriv.dat
@@ -1,0 +1,7 @@
+first_deriv:-1.00000000000000000e+00 first_deriv:-1.15470053837925168e+00 
+ 0.00000000000000000e+00   5.00000000000000000e-01  
+ 1.57079632679489656e+00   1.04719755119659785e+00  
+ 0.00000000000000000e+00   1.38888888888888881e-02   2.77777777777777762e-02   4.16666666666666644e-02   5.55555555555555525e-02   6.94444444444444475e-02   8.33333333333333287e-02   9.72222222222222099e-02   1.11111111111111105e-01   1.25000000000000000e-01   1.38888888888888895e-01   1.52777777777777762e-01   1.66666666666666657e-01   1.80555555555555552e-01   1.94444444444444420e-01   2.08333333333333315e-01   2.22222222222222210e-01   2.36111111111111105e-01   2.50000000000000000e-01   2.63888888888888895e-01   2.77777777777777790e-01   2.91666666666666630e-01   3.05555555555555525e-01   3.19444444444444420e-01   3.33333333333333315e-01   3.47222222222222210e-01   3.61111111111111105e-01   3.75000000000000000e-01   3.88888888888888840e-01   4.02777777777777735e-01   4.16666666666666630e-01   4.30555555555555525e-01   4.44444444444444420e-01   4.58333333333333315e-01   4.72222222222222210e-01   4.86111111111111105e-01   5.00000000000000000e-01  
+ 1.57079632679489656e+00   1.55691184868293275e+00   1.54303360701525194e+00   1.52915772412750273e+00   1.51528032235533483e+00   1.50139752403439686e+00   1.48750545150033808e+00   1.47360022708880711e+00   1.45967797313545344e+00   1.44573481197592613e+00   1.43176686594587399e+00   1.41777025738094609e+00   1.40374110861679124e+00   1.38967554198905918e+00   1.37556967983339828e+00   1.36141964448545805e+00   1.34722155828088708e+00   1.33297154355533487e+00   1.31866572264445048e+00   1.30430021788388251e+00   1.28987115160928023e+00   1.27537464615629270e+00   1.26080682386056875e+00   1.24616380705775787e+00   1.23144171808350888e+00   1.21663667927347063e+00   1.20174481296329239e+00   1.18676224148862319e+00   1.17168508718511211e+00   1.15650947238840818e+00   1.14123151943416024e+00   1.12584735065801733e+00   1.11035308839562874e+00   1.09474485498264351e+00   1.07901877275471048e+00   1.06317096404747891e+00   1.04719755119659785e+00  
+-1.00000000000000000e+00  -9.99411380094997748e-01  -9.99101952023264350e-01  -9.99071715784799697e-01  -9.99320671379603787e-01  -9.99848818807676620e-01  -1.00065615806901831e+00  -1.00174268916362874e+00  -1.00310841209150792e+00  -1.00475332685265606e+00  -1.00667743344707272e+00  -1.00888073187475835e+00  -1.01136322213571272e+00  -1.01412490422993584e+00  -1.01716577815742770e+00  -1.02048584391818853e+00  -1.02408510151221788e+00  -1.02796355093951619e+00  -1.03212119220008325e+00  -1.03655802529391905e+00  -1.04127405022102359e+00  -1.04626926698139711e+00  -1.05154367557503914e+00  -1.05709727600195014e+00  -1.06293006826212988e+00  -1.06904205235557836e+00  -1.07543322828229559e+00  -1.08210359604228179e+00  -1.08905315563553651e+00  -1.09628190706206019e+00  -1.10378985032185262e+00  -1.11157698541491379e+00  -1.11964331234124370e+00  -1.12798883110084258e+00  -1.13661354169370998e+00  -1.14551744411984635e+00  -1.15470053837925146e+00  
+ 5.24315391578369372e-02   3.23297271624847227e-02   1.22279151671325081e-02  -7.87389682821970638e-03  -2.79757088235719209e-02  -4.80775208189241354e-02  -6.81793328142763499e-02  -8.82811448096285367e-02  -1.08382956804980779e-01  -1.28484768800332994e-01  -1.48586580795685208e-01  -1.68688392791037367e-01  -1.88790204786389637e-01  -2.08892016781741852e-01  -2.28993828777094011e-01  -2.49095640772446225e-01  -2.69197452767798495e-01  -2.89299264763150710e-01  -3.09401076758502924e-01  -3.29502888753855139e-01  -3.49604700749207353e-01  -3.69706512744559457e-01  -3.89808324739911671e-01  -4.09910136735263997e-01  -4.30011948730616211e-01  -4.50113760725968426e-01  -4.70215572721320640e-01  -4.90317384716672855e-01  -5.10419196712024958e-01  -5.30521008707377173e-01  -5.50622820702729387e-01  -5.70724632698081602e-01  -5.90826444693433928e-01  -6.10928256688786142e-01  -6.31030068684138357e-01  -6.51131880679490571e-01  -6.71233692674842786e-01
