+++ source/module_base/test/data/two_points_periodic.dat
@@ -1,0 +1,7 @@
+periodic periodic
+ 0.00000000000000000e+00   1.00000000000000000e+00  
+ 1.00000000000000000e+00   1.00000000000000000e+00  
+ 0.00000000000000000e+00   2.77777777777777762e-02   5.55555555555555525e-02   8.33333333333333287e-02   1.11111111111111105e-01   1.38888888888888895e-01   1.66666666666666657e-01   1.94444444444444420e-01   2.22222222222222210e-01   2.50000000000000000e-01   2.77777777777777790e-01   3.05555555555555525e-01   3.33333333333333315e-01   3.61111111111111105e-01   3.88888888888888840e-01   4.16666666666666630e-01   4.44444444444444420e-01   4.72222222222222210e-01   5.00000000000000000e-01   5.27777777777777790e-01   5.55555555555555580e-01   5.83333333333333259e-01   6.11111111111111049e-01   6.38888888888888840e-01   6.66666666666666630e-01   6.94444444444444420e-01   7.22222222222222210e-01   7.50000000000000000e-01   7.77777777777777679e-01   8.05555555555555469e-01   8.33333333333333259e-01   8.61111111111111049e-01   8.88888888888888840e-01   9.16666666666666630e-01   9.44444444444444420e-01   9.72222222222222210e-01   1.00000000000000000e+00  
+ 1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00   1.00000000000000000e+00  
+ 0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00  
+ 0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00   0.00000000000000000e+00
