+++ source/module_base/test/data/two_points_second_deriv.dat
@@ -1,0 +1,7 @@
+second_deriv:5.00000000000000000e-01  second_deriv:2.50000000000000000e-01  
+ 1.00000000000000000e+00   4.00000000000000000e+00  
+ 1.00000000000000000e+00   2.00000000000000000e+00  
+ 1.00000000000000000e+00   1.08333333333333326e+00   1.16666666666666674e+00   1.25000000000000000e+00   1.33333333333333326e+00   1.41666666666666652e+00   1.50000000000000000e+00   1.58333333333333326e+00   1.66666666666666652e+00   1.75000000000000000e+00   1.83333333333333326e+00   1.91666666666666652e+00   2.00000000000000000e+00   2.08333333333333304e+00   2.16666666666666652e+00   2.25000000000000000e+00   2.33333333333333304e+00   2.41666666666666652e+00   2.50000000000000000e+00   2.58333333333333304e+00   2.66666666666666652e+00   2.75000000000000000e+00   2.83333333333333304e+00   2.91666666666666652e+00   3.00000000000000000e+00   3.08333333333333304e+00   3.16666666666666652e+00   3.25000000000000000e+00   3.33333333333333304e+00   3.41666666666666652e+00   3.50000000000000000e+00   3.58333333333333304e+00   3.66666666666666652e+00   3.75000000000000000e+00   3.83333333333333304e+00   3.91666666666666652e+00   4.00000000000000000e+00  
+ 1.00000000000000000e+00   9.77422518004115282e-01   9.58269032921810648e-01   9.42491319444444531e-01   9.30041152263374471e-01   9.20870306069958899e-01   9.14930555555555469e-01   9.12173675411522611e-01   9.12551440329218089e-01   9.16015625000000000e-01   9.22518004115226220e-01   9.32010352366254957e-01   9.44444444444444420e-01   9.59772055041152150e-01   9.77944958847736689e-01   9.98914930555555469e-01   1.02263374485596703e+00   1.04905317644032903e+00   1.07812500000000000e+00   1.10980099022633705e+00   1.14403292181069949e+00   1.18077256944444420e+00   1.21997170781892983e+00   1.26158211162551437e+00   1.30555555555555558e+00   1.35184381430041123e+00   1.40039866255144019e+00   1.45117187500000000e+00   1.50411522633744821e+00   1.55918049125514369e+00   1.61631944444444442e+00   1.67548386059670751e+00   1.73662551440329205e+00   1.79969618055555580e+00   1.86464763374485565e+00   1.93143164866255113e+00   2.00000000000000000e+00  
+-2.91666666666666685e-01  -2.50289351851851916e-01  -2.09490740740740727e-01  -1.69270833333333343e-01  -1.29629629629629706e-01  -9.05671296296297335e-02  -5.20833333333333703e-02  -1.41782407407408349e-02   2.31481481481480254e-02   5.98958333333332801e-02   9.60648148148147418e-02   1.31655092592592504e-01   1.66666666666666630e-01   2.01099537037036841e-01   2.34953703703703554e-01   2.68229166666666630e-01   3.00925925925925764e-01   3.33043981481481399e-01   3.64583333333333259e-01   3.95543981481481344e-01   4.25925925925925930e-01   4.55729166666666741e-01   4.84953703703703665e-01   5.13599537037036979e-01   5.41666666666666741e-01   5.69155092592592338e-01   5.96064814814814659e-01   6.22395833333333259e-01   6.48148148148147918e-01   6.73321759259259189e-01   6.97916666666666630e-01   7.21932870370370239e-01   7.45370370370370239e-01   7.68229166666666630e-01   7.90509259259259189e-01   8.12210648148148140e-01   8.33333333333333370e-01  
+ 4.99999999999999944e-01   4.93055555555555525e-01   4.86111111111111049e-01   4.79166666666666630e-01   4.72222222222222210e-01   4.65277777777777790e-01   4.58333333333333315e-01   4.51388888888888895e-01   4.44444444444444475e-01   4.37500000000000000e-01   4.30555555555555580e-01   4.23611111111111160e-01   4.16666666666666685e-01   4.09722222222222265e-01   4.02777777777777846e-01   3.95833333333333370e-01   3.88888888888888951e-01   3.81944444444444531e-01   3.75000000000000056e-01   3.68055555555555636e-01   3.61111111111111160e-01   3.54166666666666741e-01   3.47222222222222321e-01   3.40277777777777901e-01   3.33333333333333426e-01   3.26388888888889006e-01   3.19444444444444531e-01   3.12500000000000111e-01   3.05555555555555691e-01   2.98611111111111271e-01   2.91666666666666796e-01   2.84722222222222376e-01   2.77777777777777957e-01   2.70833333333333481e-01   2.63888888888889062e-01   2.56944444444444642e-01   2.50000000000000167e-01
